use int64 type for claims with timestamps

+++ token/jwt/claims_jwt_test.go
@@ -48,11 +48,11 @@ var jwtClaims = &JWTClaims{
 
 var jwtClaimsMap = map[string]interface{}{
 	"sub": jwtClaims.Subject,
-	"iat": float64(jwtClaims.IssuedAt.Unix()),
+	"iat": jwtClaims.IssuedAt.Unix(),
 	"iss": jwtClaims.Issuer,
-	"nbf": float64(jwtClaims.NotBefore.Unix()),
+	"nbf": jwtClaims.NotBefore.Unix(),
 	"aud": jwtClaims.Audience,
-	"exp": float64(jwtClaims.ExpiresAt.Unix()),
+	"exp": jwtClaims.ExpiresAt.Unix(),
 	"jti": jwtClaims.JTI,
 	"scp": []string{"email", "offline"},
 	"foo": jwtClaims.Extra["foo"],
